requirements.txt
wandb
bs4
langchain-text-splitters
lxml
transformers
sqlalchemy
fastapi
pg8000
uvicorn
python-dotenv
